requirements.txt
discord.py
transformers
torch
accelerate
python-dotenv
